feat; inicio da implementacao do sorteio de parnaiba
