+++ sorteio/templates/sorteio/app/sidebar.html
@@ -21,6 +21,7 @@
                     <ul class="nav child_menu">
                         <li><a href="{% url 'sorteio:sorteio_form' %}">Aleatório</a></li>
                         <li><a href="{% url 'sorteio:sorteio_bloco_periodo' %}">Por bloco e período</a></li>
+                        <li><a href="{% url 'sorteio:sorteio_parnaiba' %}">Parnaíba (Em desenvolvimento)</a></li>
                     </ul>
                 </li>
