+++ sorteio/templates/sorteio/sorteio_bloco_parnaiba.html
@@ -1,0 +1,93 @@
+{% extends "sorteio/app/base_site.html" %}
+
+{% block title %} Table Dynamic {% endblock title %}
+
+{% block stylesheets %}
+{{ block.super }}
+<link href="/static/vendors/datatables.net-bs/css/dataTables.bootstrap.min.css" rel="stylesheet">
+<link href="/static/vendors/datatables.net-buttons-bs/css/buttons.bootstrap.min.css" rel="stylesheet">
+<link href="/static/vendors/datatables.net-fixedheader-bs/css/fixedHeader.bootstrap.min.css" rel="stylesheet">
+<link href="/static/vendors/datatables.net-responsive-bs/css/responsive.bootstrap.min.css" rel="stylesheet">
+<link href="/static/vendors/datatables.net-scroller-bs/css/scroller.bootstrap.min.css" rel="stylesheet">
+{% endblock stylesheets %}
+
+{% block content %}
+<div class="right_col" role="main"
+  style="display: flex; flex-direction: row; justify-content: center; align-items: flex-start;">
+  <div class="info full-width" style="padding-right: 32px;">
+
+    {% if sorteios %}
+    <table id="datatable-buttons" class="table table-striped table-bordered">
+      <tbody>
+        {% for sorteio in sorteios %}
+        <tr>
+          <td>{{ forloop.counter }} </a></td>
+          <td>{{ sorteio.defensor }}</td>
+          <td>{{ sorteio.data|date:"d M" }}</td>
+          <td>{{ sorteio.data|date:"l" }}</td>
+        </tr>
+        {% empty %}
+        {% endfor %}
+      </tbody>
+      <thead>
+        <tr>
+          <th>Ordem</th>
+          <th>Defensor(a)</th>
+          <th>Data</th>
+          <th>Dia da Semana</th>
+        </tr>
+      </thead>
+    </table>
+    {% else %}
+
+    <h1>Sorteio - Parnaíba</h1>
+    <h2>Em desenvolvimento</h2>
+
+    <hr>
+
+    <p>Selecione os defensores que irão participar do sorteio no recesso desse ano</p>
+
+    {% endif %}
+
+  </div>
+  <div class="full-width">
+    <form method="POST">
+      {% csrf_token %}
+      {{ form.as_p }}
+      <input type="submit" value="Sortear" id="btn-sortear" style="padding: 8px 16px;" class="full-width"
+        onclick="mostrarMensagem()" /><br>
+      <div id="mensagem" style="text-align: center; display: none;">
+        <h1>Sorteio em execução</h1>
+        <h2>Aguarde, esse processo pode levar alguns minutos!</h2>
+      </div>
+    </form>
+  </div>
+</div>
+
+<script>
+  var messagem = document.getElementById('mensagem');
+  function mostrarMensagem() {
+    mensagem.style.display = 'block';
+  }
+</script>
+{% endblock content %}
+
+{% block javascripts %}
+{{ block.super }}
+<!-- Datatables -->
+<script src="/static/vendors/datatables.net/js/jquery.dataTables.min.js"></script>
+<script src="/static/vendors/datatables.net-bs/js/dataTables.bootstrap.min.js"></script>
+<script src="/static/vendors/datatables.net-buttons/js/dataTables.buttons.min.js"></script>
+<script src="/static/vendors/datatables.net-buttons-bs/js/buttons.bootstrap.min.js"></script>
+<script src="/static/vendors/datatables.net-buttons/js/buttons.flash.min.js"></script>
+<script src="/static/vendors/datatables.net-buttons/js/buttons.html5.min.js"></script>
+<script src="/static/vendors/datatables.net-buttons/js/buttons.print.min.js"></script>
+<script src="/static/vendors/datatables.net-fixedheader/js/dataTables.fixedHeader.min.js"></script>
+<script src="/static/vendors/datatables.net-keytable/js/dataTables.keyTable.min.js"></script>
+<script src="/static/vendors/datatables.net-responsive/js/dataTables.responsive.min.js"></script>
+<script src="/static/vendors/datatables.net-responsive-bs/js/responsive.bootstrap.js"></script>
+<script src="/static/vendors/datatables.net-scroller/js/dataTables.scroller.min.js"></script>
+<script src="/static/vendors/jszip/dist/jszip.min.js"></script>
+<script src="/static/vendors/pdfmake/build/pdfmake.min.js"></script>
+<script src="/static/vendors/pdfmake/build/vfs_fonts.js"></script>
+{% endblock javascripts %}
